requirements.txt
aiohttp==3.7.3
aiomysql==0.0.21
aioredis==1.3.1
cryptography==3.4.4
fastapi==0.63.0
gunicorn==20.0.4
httptools==0.1.1
pytz==2020.5
PyYAML==5.4.1
sentry-sdk==0.20.0
tortoise-orm==0.16.21
ujson==4.0.2
uvicorn==0.13.3
uvloop==0.15.0
